pyproject.toml
[project]
name = "trustmed"
version = "0.1.0"
description = ""
authors = [
    {name = "cpuoverload"}
]
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "chromadb (>=0.6.3,<0.7.0)",
    "llama-index (>=0.12.18,<0.13.0)",
    "llama-index-llms-ollama (>=0.5.2,<0.6.0)",
    "llama-index-embeddings-ollama (>=0.5.0,<0.6.0)",
    "llama-index-vector-stores-chroma (>=0.4.1,<0.5.0)",
    "streamlit (>=1.42.0,<2.0.0)",
    "spacy (>=3.8.4,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "llama-index-llms-dashscope (>=0.3.0,<0.4.0)",
    "llama-index-embeddings-dashscope (>=0.3.0,<0.4.0)",
    "ragas (>=0.2.13,<0.3.0)",
    "rapidfuzz (>=3.12.1,<4.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
